+++ src/pages/Calendar.tsx
@@ -3,10 +3,11 @@ import { useSetRecoilState } from 'recoil';
 import { useNavigate } from 'react-router';
 import { getSchedule } from '@/lib/api';
 import { Schedule, AlertState } from '@/lib/types';
-import { CALENDAR_TEXTS } from '@/constants/calendar';
 import { alertState } from '@/states/stateAlert';
+import CalendarHeader from '@/components/Calendar/CalendarHeader';
 import CalendarBody from '@/components/Calendar/CalendarBody';
 import CalendarList from '@/components/Calendar/CalendarList';
+import ToggleButton from '@/components/Calendar/ToggleButton';
 import Loading from '@/components/Loading';
 import Alert from '@/components/Alert';
 import dayjs from 'dayjs';
@@ -25,7 +26,7 @@ const Calendar = () => {
   const navigate = useNavigate();
   const weekDays = ['일', '월', '화', '수', '목', '금', '토'];
 
-  const fetchData = async () => {
+  const getScheduleData = async () => {
     try {
       setIsLoading(true);
       const res = await getSchedule();
@@ -44,90 +45,30 @@ const Calendar = () => {
   };
 
   useEffect(() => {
-    !localStorage.getItem('authToken') && navigate('/login');
-    if (localStorage.getItem('authToken')) {
-      fetchData();
-    }
+    localStorage.getItem('authToken') ? getScheduleData() : navigate('/login');
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-
-  const prevMonth = () => {
-    setCurrentMonth(prevMonth => prevMonth.subtract(1, 'month'));
-  };
-
-  const nextMonth = () => {
-    setCurrentMonth(prevMonth => prevMonth.add(1, 'month'));
-  };
-
-  const handleClickToggle = () => {
-    setToggleButton(!toggleButton);
-  };
-
-  const handleClickToday = () => {
-    setCurrentMonth(dayjs(dayjs().format('YYYY-MM-DD')));
-  };
-
-  const handleClickDuty = () => {
-    setDutyActive(!dutyActive);
-  };
-
-  const handleClickAnnual = () => {
-    setAnnualActive(!annualActive);
-  };
 
   return (
     <Container>
       {isLoading && <Loading />}
       <Alert />
-      <ToggleButton>
-        <Button
-          className={toggleButton ? 'calendar-button active' : 'calendar-button'}
-          onClick={handleClickToggle}
-          disabled={!scheduleData}
-        >
-          {CALENDAR_TEXTS.calendar}
-        </Button>
-        <Button
-          className={toggleButton ? 'list-button' : 'list-button active'}
-          onClick={handleClickToggle}
-          disabled={!scheduleData}
-        >
-          {CALENDAR_TEXTS.list}
-        </Button>
-      </ToggleButton>
-      {toggleButton ? (
-        <>
-          <button className="today-button" onClick={handleClickToday} disabled={!scheduleData}>
-            {CALENDAR_TEXTS.today}
-          </button>
-          <FilterButtons>
-            <button
-              className={dutyActive ? 'duty-button active' : 'duty-button'}
-              onClick={handleClickDuty}
-              disabled={!scheduleData}
-            >
-              {CALENDAR_TEXTS.duty}
-            </button>
-            <button
-              className={annualActive ? 'annual-button active' : 'annual-button'}
-              onClick={handleClickAnnual}
-              disabled={!scheduleData}
-            >
-              {CALENDAR_TEXTS.annual}
-            </button>
-          </FilterButtons>
-          <Header>
-            <CalendarButtons>
-              <button className="prev-button" onClick={prevMonth} disabled={!scheduleData}>
-                &lt;
-              </button>
-              <button className="next-button" onClick={nextMonth} disabled={!scheduleData}>
-                &gt;
-              </button>
-              <MonthWrapper>{currentMonth.format('YYYY년 M월')}</MonthWrapper>
-            </CalendarButtons>
-          </Header>
-          {scheduleData ? (
+      <WrapHeader>
+        {toggleButton && (
+          <CalendarHeader
+            currentMonth={currentMonth}
+            setCurrentMonth={setCurrentMonth}
+            dutyActive={dutyActive}
+            annualActive={annualActive}
+            setDutyActive={setDutyActive}
+            setAnnualActive={setAnnualActive}
+          />
+        )}
+        <ToggleButton toggleButton={toggleButton} setToggleButton={setToggleButton} />
+      </WrapHeader>
+      <WrapBody>
+        {scheduleData &&
+          (toggleButton ? (
             <>
               <Weeks>
                 {weekDays.map(day => (
@@ -144,17 +85,9 @@ const Calendar = () => {
               />
             </>
           ) : (
-            <>
-              <Weeks></Weeks>
-              <WeekLoading></WeekLoading>
-            </>
-          )}
-        </>
-      ) : scheduleData ? (
-        <CalendarList scheduleData={scheduleData} />
-      ) : (
-        ''
-      )}
+            <CalendarList scheduleData={scheduleData} />
+          ))}
+      </WrapBody>
     </Container>
   );
 };
@@ -166,8 +99,10 @@ const Container = styled.div`
   width: 100%;
   min-width: 1100px;
   height: calc(100% - 64px);
-  padding: 0 70px 40px 70px;
+  padding: 0 70px 20px 70px;
   box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
 
   .today-button {
     position: absolute;
@@ -184,119 +119,17 @@ const Container = styled.div`
   }
 `;
 
-const WeekLoading = styled.div`
+const WrapHeader = styled.div`
+  display: flex;
   width: 100%;
-  height: calc(100% - 90px);
-  border-right: 1px solid ${props => props.theme.gray};
-  border-left: 1px solid ${props => props.theme.gray};
-  border-bottom: 1px solid ${props => props.theme.gray};
-  border-bottom-right-radius: 8px;
-  border-bottom-left-radius: 8px;
-  background-color: ${props => props.theme.white};
-  box-sizing: border-box;
+  gap: 30px;
 `;
 
-const ToggleButton = styled.div`
-  position: absolute;
-  right: 70px;
-`;
-
-const Button = styled.button`
-  height: 30px;
-  border: none;
-  outline: none;
-  border: 1px solid ${props => props.theme.gray};
-  background-color: ${props => props.theme.white};
-  color: ${props => props.theme.gray};
-  cursor: pointer;
-
-  &.calendar-button {
-    width: 45px;
-    border-top-left-radius: 4px;
-    border-bottom-left-radius: 4px;
-    border-right: none;
-  }
-  &.list-button {
-    width: 45px;
-    border-top-right-radius: 4px;
-    border-bottom-right-radius: 4px;
-  }
-  &.active {
-    border: 2px solid ${props => props.theme.primary};
-    background-color: rgba(37, 64, 135, 0.3);
-    color: ${props => props.theme.primary};
-  }
-`;
-
-const FilterButtons = styled.div`
-  position: absolute;
-  right: 0px;
+const WrapBody = styled.div`
   display: flex;
-  margin-right: 190px;
-  gap: 5px;
-
-  button {
-    width: 45px;
-    height: 30px;
-    outline: none;
-    border-radius: 4px;
-    border: 1px solid ${props => props.theme.gray};
-    background-color: ${props => props.theme.white};
-    color: ${props => props.theme.gray};
-    cursor: pointer;
-
-    &.active {
-      background-color: ${props => props.theme.primary};
-      color: ${props => props.theme.white};
-    }
-  }
-`;
-
-const Header = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
+  flex-direction: column;
   width: 100%;
-  height: 30px;
-
-  .prev-button,
-  .next-button,
-  .calendar-button,
-  .list-button {
-    height: 30px;
-    border: none;
-    outline: none;
-    border: 1px solid ${props => props.theme.gray};
-    background-color: ${props => props.theme.white};
-    color: ${props => props.theme.gray};
-    cursor: pointer;
-  }
-
-  .calendar-button,
-  .prev-button {
-    width: 45px;
-    border-top-left-radius: 4px;
-    border-bottom-left-radius: 4px;
-    border-right: none;
-  }
-
-  .list-button,
-  .next-button {
-    width: 45px;
-    border-top-right-radius: 4px;
-    border-bottom-right-radius: 4px;
-  }
-`;
-
-const CalendarButtons = styled.div`
-  display: flex;
-`;
-
-const MonthWrapper = styled.div`
-  margin-left: 70px;
-  font-size: 24px;
-  font-weight: bold;
-  color: ${props => props.theme.primary};
+  height: 100%;
 `;
 
 const Weeks = styled.div`
